added first translation of images
needs fix of algorithm, maybe needs performance boost

+++ src/System.Svg.Render.EPL/SvgTextBaseTranslator.cs
@@ -47,13 +47,13 @@ namespace System.Svg.Render.EPL
                                          out x,
                                          out y);
 
-      var fontSizeVector = new PointF(fontSize * -1,
+      var fontSizeVector = new PointF(fontSize * -1f,
                                       0f);
       this.SvgUnitCalculator.ApplyMatrix(fontSizeVector,
                                          matrix,
                                          out fontSizeVector);
       fontSize = this.SvgUnitCalculator.GetLengthOfVector(fontSizeVector);
-      var rotationTranslation = this.SvgUnitCalculator.GetRotationTranslation(fontSizeVector);
+      var rotationTranslation = (int) this.SvgUnitCalculator.GetRotationTranslation(fontSizeVector);
 
       object fontSelection;
       object multiplier;
